+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/directconnectivity/rntbd/RntbdLoopNIO.java
@@ -4,7 +4,8 @@
 package com.azure.cosmos.implementation.directconnectivity.rntbd;
 
 import io.netty.channel.EventLoopGroup;
-import io.netty.channel.nio.NioEventLoopGroup;
+import io.netty.channel.MultiThreadIoEventLoopGroup;
+import io.netty.channel.nio.NioIoHandler;
 import io.netty.channel.socket.SocketChannel;
 import io.netty.channel.socket.nio.NioSocketChannel;
 
@@ -19,7 +20,7 @@ public class RntbdLoopNIO implements RntbdLoop {
 
     @Override
     public EventLoopGroup newEventLoopGroup(int threads, ThreadFactory threadFactory) {
-        return new NioEventLoopGroup(threads, threadFactory);
+        return new MultiThreadIoEventLoopGroup(threads, threadFactory, NioIoHandler.newFactory());
     }
 
     @Override